+++ public/hike/asset_hike_map.json
@@ -1,0 +1,6 @@
+{
+  "details": [
+    []
+  ],
+  "status": 200
+}
