+++ public/hike/tour/details/06a5e8td-11a4-41r7-bbua-0696bane.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24425,
+    "title": "Assign Data to Segment",
+    "created": 1585629194,
+    "updated": 1615281650,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24425/download/17804",
+    "fileSize": "22.53 KB",
+    "fileName": "assigndatatothesegmentv9.zip",
+    "fid": 17804,
+    "checksum": "d2f5decde0b6b4b094a188eecf98e3317137e639db488d17c35ceb8896ca760e",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/assign-data-segment-0",
+    "alias": "hikes/tour/assign-data-segment-0",
+    "description": "Learn how to assign data to a Segment widget.",
+    "details": "<p><!--StartFragment--></p> <div> <div> <p>This learning module is a Volt MX HIKE.&nbsp; A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar.</p> <p>&nbsp;Learn how to:</p> <ul> <li>Change the UI and add new widgets to a Segment widget</li> <li>Bind back-end services to the UI elements in a Segment</li> </ul> </div> </div> <p><!--EndFragment--></p>",
+    "cards": 13,
+    "time": "10 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "06a5e8td-11a4-41r7-bbua-0696bane",
+    "hikeVersion": "1.0",
+    "weight": "5"
+  }
+}
