+++ public/hike/tour/details/06a67btd-a5a2-45rf-96u9-29d2c8n2.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24437,
+    "title": "Use Storage Services",
+    "created": 1585630177,
+    "updated": 1586945144,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24437/download/17160",
+    "fileSize": "21.1 KB",
+    "fileName": "usestorageservicev9.zip",
+    "fid": 17160,
+    "checksum": "9d6dcb2ff966f66f9ec7ce697bebed54718f52b284955ed797e3d87c1bc5d5bb",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/use-storage-services",
+    "alias": "hikes/tour/use-storage-services",
+    "description": "Create a storage services, populate, and connect the data to the form.",
+    "details": "<p><!--StartFragment--></p> <div> <div> <p>This learning module is a Volt MX HIKE.&nbsp; A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar.</p> <p>&nbsp;Learn how to:</p> <ul> <li>Create a form</li> <li>Create a storage service</li> <li>Populate data in the storage service</li> <li>Connect data to the form</li> </ul> </div> </div> <p><!--EndFragment--></p>",
+    "cards": 14,
+    "time": "15 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "06a67btd-a5a2-45rf-96u9-29d2c8n2",
+    "hikeVersion": "1.0",
+    "weight": "5"
+  }
+}
