+++ public/hike/tour/details/1574aft1-16af-4bre-90ud-1fbc7cna.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24053,
+    "title": "Dynamic UI Generation",
+    "created": 1573815016,
+    "updated": 1578655923,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24053/download/16131",
+    "fileSize": "5.68 KB",
+    "fileName": "dynamicform.zip",
+    "fid": 16131,
+    "checksum": "510d95f7b45d0196d718b457d52172464d173f3ca6e77e5fcbd2e1e7389b9265",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/dynamic-ui-generation",
+    "alias": "hikes/tour/dynamic-ui-generation",
+    "description": "Dynamic UI generation using Data panel.",
+    "details": "",
+    "cards": 3,
+    "time": "5 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "1574aft1-16af-4bre-90ud-1fbc7cna",
+    "hikeVersion": "1.0",
+    "weight": "0"
+  }
+}
