+++ public/hike/tour/details/21ea72t7-d5a6-48ra-b0u2-ba8b54n3.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24435,
+    "title": "Add an Authentication Source",
+    "created": 1585630068,
+    "updated": 1615281900,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24435/download/17806",
+    "fileSize": "32.38 KB",
+    "fileName": "userrepositoryauthenticationv9.zip",
+    "fid": 17806,
+    "checksum": "b00ace6ea699ab761a1917a34a335ae424dba95be1f61cbf6650dc77a81119ec",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/add-authentication-source-0",
+    "alias": "hikes/tour/add-authentication-source-0",
+    "description": "Add Volt MX User Repository to your app to authorize users.",
+    "details": "<p><!--StartFragment--></p> <div> <div> <p>This learning module is a Volt MX HIKE.&nbsp; A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar.</p> <p>&nbsp;Learn how to:</p> <ul> <li>Add Volt MX User Repository as an authentication source</li> <li>Add users to the repository</li> <li>Access app using the login credentials from the Volt MX User Repository</li> </ul> </div> </div> <p><!--EndFragment--></p>",
+    "cards": 19,
+    "time": "17 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "21ea72t7-d5a6-48ra-b0u2-ba8b54n3",
+    "hikeVersion": "1.0",
+    "weight": "3"
+  }
+}
