+++ public/hike/tour/details/295e73td-95a1-44r8-8cud-93e713n1.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24433,
+    "title": "Understand Identity Services",
+    "created": 1585629962,
+    "updated": 1586944996,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24433/download/17156",
+    "fileSize": "21.46 KB",
+    "fileName": "understandidentityservicesv9.zip",
+    "fid": 17156,
+    "checksum": "41ff3309c59a43ba831b05738aa9fadcaff7d67f9f4838406a9d081cfcfc8005",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/understand-identity-services-0",
+    "alias": "hikes/tour/understand-identity-services-0",
+    "description": "Add authentication to your app.",
+    "details": "<p><!--StartFragment--></p> <div> <div> <p>This learning module is a Volt MX HIKE.&nbsp; A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar.</p> <p>&nbsp;Learn how to secure your application by using an authentication mechanism.</p> </div> </div> <p><!--EndFragment--></p>",
+    "cards": 13,
+    "time": "15 MIns",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "295e73td-95a1-44r8-8cud-93e713n1",
+    "hikeVersion": "1.0",
+    "weight": "1"
+  }
+}
