+++ public/hike/tour/details/3b490at5-80ae-46rb-a4ua-3cd903n7.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24428,
+    "title": "Creating a Component",
+    "created": 1585629363,
+    "updated": 1586944337,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24428/download/17151",
+    "fileSize": "52.67 KB",
+    "fileName": "createemailidcomponentv9.zip",
+    "fid": 17151,
+    "checksum": "099f110c447aef3dd0037a8a59e215e7d7c7b570763725f8511723110aa19303",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/creating-component-0",
+    "alias": "hikes/tour/creating-component-0",
+    "description": "Learn how to create a component.",
+    "details": "<p><!--StartFragment--></p> <div> <div> <p>This learning module is a Volt MX HIKE.&nbsp; A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar.</p> <p>&nbsp;Learn how to:</p> <ul> <li>Create a login form</li> <li>Create a layout for the component</li> <li>Add data validations to widgets</li> <li>Add the component to the login form</li> </ul> </div> </div> <p><!--EndFragment--></p>",
+    "cards": 30,
+    "time": "22 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "3b490at5-80ae-46rb-a4ua-3cd903n7",
+    "hikeVersion": "1.0",
+    "weight": "8"
+  }
+}
