+++ public/hike/tour/details/7408efte-e2aa-4cr8-a9ub-d7ba36n8.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24825,
+    "title": "Launch your First Mobile App",
+    "created": 1608540013,
+    "updated": 1615279231,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24825/download/17790",
+    "fileSize": "20.45 KB",
+    "fileName": "launchyourfirstmobileappv92.zip",
+    "fid": 17790,
+    "checksum": "629127d7f68f964d2827da15d82225e4fbcaae5f72dc85b7b6b843ac3feb34b9",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/launch-your-first-mobile-app-1",
+    "alias": "hikes/tour/launch-your-first-mobile-app-1",
+    "description": "Open a native mobile app and preview it on your mobile device.",
+    "details": "<p><!--StartFragment--></p> <p><br /> This learning module is a Volt MX HIKE. &nbsp;A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar:&nbsp;</p> <p> &nbsp;</p> <p>Learn how to:&nbsp;</p> <p>&nbsp; &nbsp; Import a sample native app that shows lists of Accounts and Recipients&nbsp;</p> <p>&nbsp; &nbsp; Preview the sample app on a mobile device&nbsp;</p> <p><!--EndFragment--></p>",
+    "cards": 12,
+    "time": "11 Mins",
+    "platformVersion": "Iris 9.2",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "7408efte-e2aa-4cr8-a9ub-d7ba36n8",
+    "hikeVersion": "1.0",
+    "weight": "1"
+  }
+}
