+++ public/hike/tour/details/7a7e14td-bda1-4fr1-bbu6-26a503nb.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24430,
+    "title": "Introduction to Volt MX JavaScript",
+    "created": 1585629482,
+    "updated": 1586944403,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24430/download/17153",
+    "fileSize": "25.72 KB",
+    "fileName": "introductiontokonyjavascriptv9.zip",
+    "fid": 17153,
+    "checksum": "b88a78e7da20bb4020f338121800c0394497b072744a21eabb32e3544eb09295",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/introduction-quantum-javascript",
+    "alias": "hikes/tour/introduction-quantum-javascript",
+    "description": "Understanding the basics of Volt MX JavaScript.",
+    "details": "<p><!--StartFragment--></p> <div> <div> <p>This learning module is a Volt MX HIKE.&nbsp; A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar.</p> <p>&nbsp;Learn how to:</p> <ul> <li>Use JavaScript to create a login form</li> <li>Get insights on the validate and onNavigate functions</li> </ul> </div> </div> <p><!--EndFragment--></p>",
+    "cards": 16,
+    "time": "20 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "7a7e14td-bda1-4fr1-bbu6-26a503nb",
+    "hikeVersion": "1.0",
+    "weight": "10"
+  }
+}
