+++ public/hike/tour/details/9167adt8-36ac-4crc-90ua-da60c8n7.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24827,
+    "title": "Introduction to Internationalization",
+    "created": 1608540137,
+    "updated": 1608540137,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24827/download/17647",
+    "fileSize": "32.3 KB",
+    "fileName": "introductiontointernationalization92.zip",
+    "fid": 17647,
+    "checksum": "2688c6c318c387cfb29cd654eb5aac822de52a58ca3dd270c51b3927e92f865c",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/introduction-internationalization-1",
+    "alias": "hikes/tour/introduction-internationalization-1",
+    "description": "Learn the basics of Internationalization(i18n).",
+    "details": "<p><!--StartFragment--></p> <p><!--StartFragment--></p> <p>This learning module is a Volt MX Hike.&nbsp;A Hike is an interactive tour in Volt MX Iris that guides you through a step-by-step process of each module. You can access the Hikes catalog by clicking the Hikes icon in the Iris toolbar.</p> <p>Learn how to:</p> <ul> <li>Configure an app for different locales</li> <li>View the locale changes for the app on a device</li> </ul> <p><!--EndFragment--></p> <p><!--EndFragment--></p>",
+    "cards": 19,
+    "time": "17 Mins",
+    "platformVersion": "Iris 9.2",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "9167adt8-36ac-4crc-90ua-da60c8n7",
+    "hikeVersion": "1.0",
+    "weight": "11"
+  }
+}
