+++ public/hike/tour/details/953f7at8-eba4-40rd-86uc-bf1f0dnd.json
@@ -1,0 +1,27 @@
+{
+  "tourDetails": {
+    "nid": 22914,
+    "title": "1. New Iris UI",
+    "created": 1546833820,
+    "updated": 1578464532,
+    "fileURL": 17243,
+    "fileSize": "8.79 KB",
+    "fileName": "voltmxirisintro.zip",
+    "fid": 17243,
+    "checksum": "c23ff55adb515a0eb7b400f2e23ba2300b31cae0fafb528c7a5896f70de7c7aa",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/1-import-and-launch-web-app",
+    "alias": "hikes/tour/1-new-ui",
+    "description": "HCL Volt MX Iris User Interface Designs",
+    "details": "<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamcos laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. <img alt=\"\" src=\"/volt-mx-tutorials/contents/1-volt-mx-changes/assets/iris-desktop.png\" style=\"height:759px; width:1423px\" /></p>",
+    "cards": 7,
+    "time": "8 Mins",
+    "platformVersion": "Iris 9.2.3",
+    "category": [
+      "Volt MX Changes"
+    ],
+    "image": "/default/hike-default.png",
+    "kuid": "953f7at8-eba4-40rd-86uc-bf1f0dnd",
+    "hikeVersion": "1.0",
+    "weight": "1"
+  }
+}
