+++ public/hike/tour/details/9b0f07t4-e9a8-49rb-83u2-876264n4.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24424,
+    "title": "Introduction to Segment",
+    "created": 1585629142,
+    "updated": 1615281581,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24424/download/17803",
+    "fileSize": "32.51 KB",
+    "fileName": "introductiontosegmentv9.zip",
+    "fid": 17803,
+    "checksum": "7b5d27b81126c58fcc507a663059589bfba37fe42d75585503de0723cb29cc7d",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/introduction-segment-0",
+    "alias": "hikes/tour/introduction-segment-0",
+    "description": "Learn the basics of the Segment widget.",
+    "details": "<p><!--StartFragment--></p> <div> <div> <p>This learning module is a Volt MX HIKE.&nbsp; A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar.</p> <p>&nbsp;Learn how to:</p> <ul> <li>Create a FlexForm with a Segment widget</li> <li>Add widgets to a Segment widget</li> <li>Assign sample data to a Segment widget</li> </ul> </div> </div> <p><!--EndFragment--></p>",
+    "cards": 20,
+    "time": "15 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "9b0f07t4-e9a8-49rb-83u2-876264n4",
+    "hikeVersion": "1.0",
+    "weight": "4"
+  }
+}
