+++ public/hike/tour/details/a617eft5-7ca8-4ard-8eu8-94cdean5.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24403,
+    "title": "Create UI from Collections",
+    "created": 1585625754,
+    "updated": 1586774913,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24403/download/17118",
+    "fileSize": "30.95 KB",
+    "fileName": "createuifromcollectionsmob_v9.zip",
+    "fid": 17118,
+    "checksum": "ff455d26cd6bb9a363e1e03e0537ad7e41721ecb7c990f60a4797228f711ee47",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/create-ui-collections",
+    "alias": "hikes/tour/create-ui-collections",
+    "description": "Create the user interface for an app with Volt MX Iris",
+    "details": "<p><!--StartFragment--></p> <div> <div> <p><!--StartFragment--></p> <div> <div> <p>This learning module is a Volt MX HIKE.  A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar:&nbsp;</p> <p> &nbsp;</p> <p>Learn how to:&nbsp;</p> <ul> <li>Create forms with UI elements for user interaction</li> <li>Add navigation to access the different forms in this app</li> </ul> </div> </div> </div> </div> <p><!--EndFragment--></p>",
+    "cards": 19,
+    "time": "20 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "a617eft5-7ca8-4ard-8eu8-94cdean5",
+    "hikeVersion": "1.0",
+    "weight": "2"
+  }
+}
