+++ public/hike/tour/details/c108beta-87ac-46re-87u3-92e61cnf.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24419,
+    "title": "Add Edit Recipient",
+    "created": 1585628395,
+    "updated": 1615281000,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24419/download/17799",
+    "fileSize": "60.11 KB",
+    "fileName": "editpayeeweb_v9.zip",
+    "fid": 17799,
+    "checksum": "43aa52960981158032c73b6de9dab1348d080fff394f32ecefe1567a92e2e5a0",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/add-edit-recipient-0",
+    "alias": "hikes/tour/add-edit-recipient-0",
+    "description": "Add the capability to edit the details of a recipient.",
+    "details": "<p><!--StartFragment--></p> <div> <div> <p>This learning module is a Volt MX HIKE.  A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar:&nbsp;</p> <p> &nbsp;</p> <p>Learn how to:&nbsp;</p> <ul> <li>Generate a screen for editing recipient details</li> <li>Design the UI for the Edit Recipient form</li> <li>Connect the data from the services to the UI elements</li> </ul> </div> </div> <p><!--EndFragment--></p>",
+    "cards": 34,
+    "time": "35 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "c108beta-87ac-46re-87u3-92e61cnf",
+    "hikeVersion": "1.0",
+    "weight": "8"
+  }
+}
