+++ public/hike/tour/details/c99652td-06a5-43ra-87u4-9e8895n7.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24426,
+    "title": "Introduction to Segment Templates",
+    "created": 1585629249,
+    "updated": 1615281715,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24426/download/17805",
+    "fileSize": "26 KB",
+    "fileName": "introductiontosegmenttemplatesv9.zip",
+    "fid": 17805,
+    "checksum": "3c063f185ef1bb63740362788b6469052062446a46e9703cbc5a8f8abae824b4",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/introduction-segment-templates-0",
+    "alias": "hikes/tour/introduction-segment-templates-0",
+    "description": "Know more about using Segment Templates.",
+    "details": "<p><!--StartFragment--></p> <div> <div> <p>This learning module is a Volt MX HIKE.&nbsp; A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar.</p> <p>&nbsp;Learn how to:</p> <ul> <li>Convert a row in a Segment into a template.</li> <li>Add sample data to widgets in a Segment using template.</li> </ul> </div> </div> <p><!--EndFragment--></p>",
+    "cards": 15,
+    "time": "12 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "c99652td-06a5-43ra-87u4-9e8895n7",
+    "hikeVersion": "1.0",
+    "weight": "6"
+  }
+}
