+++ public/hike/tour/details/d75f6et0-c0af-4dr3-bbue-03c122nd.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24436,
+    "title": "Use Sample Services",
+    "created": 1585630127,
+    "updated": 1586945117,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24436/download/17159",
+    "fileSize": "17.39 KB",
+    "fileName": "usesampleservicesv9.zip",
+    "fid": 17159,
+    "checksum": "ce6738fb4a845df8adf94c1ec585fcbcf61e1de48f27e428df0f57aafaf638a7",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/use-sample-services-0",
+    "alias": "hikes/tour/use-sample-services-0",
+    "description": "Connect a sample back-end database to a form",
+    "details": "<p><!--StartFragment--></p> <div> <div> <p>This learning module is a Volt MX HIKE.&nbsp; A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar.</p> <p>&nbsp;Learn how to:</p> <ul> <li>Create a form</li> <li>Connect to a sample back-end database</li> <li>Add data to the form. You can follow the same steps to bind your own back-end data to your app.</li> </ul> </div> </div> <p><!--EndFragment--></p>",
+    "cards": 11,
+    "time": "12 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "d75f6et0-c0af-4dr3-bbue-03c122nd",
+    "hikeVersion": "1.0",
+    "weight": "4"
+  }
+}
