+++ public/hike/tour/details/ec5ceatd-50a0-4fre-b3u7-c3e41cn2.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24410,
+    "title": "Publish and Share the App",
+    "created": 1585626952,
+    "updated": 1586775145,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24410/download/17125",
+    "fileSize": "13.57 KB",
+    "fileName": "shareandpreviewappmobilev9.zip",
+    "fid": 17125,
+    "checksum": "868bad9ce328ec89e10ceba0eb74d3ab09e6e0e3afd974ac18ed80223ad88184",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/publish-and-share-app",
+    "alias": "hikes/tour/publish-and-share-app",
+    "description": "Use the Volt MX App to preview the Recipients app on a mobile device.",
+    "details": "<p><!--StartFragment--></p> <p>This learning module is a Volt MX HIKE.  A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar:&nbsp;</p> <p> &nbsp;</p> <p>Learn how to:&nbsp;</p> <ul> <li>Preview the Recipients app on your mobile device.</li> </ul> <p><!--EndFragment--></p>",
+    "cards": 7,
+    "time": "10 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "ec5ceatd-50a0-4fre-b3u7-c3e41cn2",
+    "hikeVersion": "1.0",
+    "weight": "9"
+  }
+}
