+++ public/hike/tour/details/f31910t6-a4a7-42rd-a4ud-07a264n3.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24420,
+    "title": "Responsive Web Breakpoints",
+    "created": 1585628476,
+    "updated": 1585918354,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24420/download/16995",
+    "fileSize": "36.34 KB",
+    "fileName": "responsivewebbreakpointsv9.zip",
+    "fid": 16995,
+    "checksum": "e0da8fb14306d982e67bf83e3645bab2a4b48b8ce80d3b1eec3302e14865936a",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/responsive-web-breakpoints",
+    "alias": "hikes/tour/responsive-web-breakpoints",
+    "description": "Make your web app responsive by using breakpoints.",
+    "details": "<p><!--StartFragment--></p> <div> <div> <p>This learning module is a Volt MX HIKE.  A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar:&nbsp;</p> <p> &nbsp;</p> <p>Learn how to:&nbsp;</p> <ul> <li>Add three breakpoints for the forms Create Recipient and Edit Recipient screens.</li> <li>See how the&nbsp;<strong>Live Preview</strong>&nbsp;reacts to breakpoints in real-time.</li> </ul> </div> </div> <p><!--EndFragment--></p>",
+    "cards": 20,
+    "time": "20 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "f31910t6-a4a7-42rd-a4ud-07a264n3",
+    "hikeVersion": "1.0",
+    "weight": "9"
+  }
+}
