+++ public/hike/tour/details/f9e5f0t5-a8a5-4arb-bfu1-50d392n4.json
@@ -1,0 +1,25 @@
+{
+  "tourDetails": {
+    "nid": 24439,
+    "title": "NFIs Using Third-party Library",
+    "created": 1585630293,
+    "updated": 1615282130,
+    "fileURL": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/24439/download/17808",
+    "fileSize": "29.83 KB",
+    "fileName": "nfibarcodev9.zip",
+    "fid": 17808,
+    "checksum": "690270167c676acce5743da5c055c56deaa0bc1f4ba14164eff2fc60506e910a",
+    "url": "https://opensource.hcltechsw.com/volt-mx-tutorials/hikes/tour/nfis-using-third-party-library",
+    "alias": "hikes/tour/nfis-using-third-party-library",
+    "description": "Use NFIs to access third party libraries, specifically Android barcode scanner.",
+    "details": "<p><!--StartFragment--></p> <div> <div> <p>This learning module is a Volt MX HIKE.&nbsp; A HIKE is an interactive tour inside Volt MX Iris that guides you through step-by-step process of each module. You can access the HIKES catalog by clicking the Hike icon in the Iris toolbar.</p> <p>Learn how to:</p> <ul> <li>Import a Collection library to an app and use it in a form</li> <li>Configure the Native Function APIs required for barcode scanner for an Android app</li> <li>Add code in the Form controller for the barcode scanning functionality</li> </ul> </div> </div> <p><!--EndFragment--></p>",
+    "cards": 19,
+    "time": "20 Mins",
+    "platformVersion": "Iris 9.0",
+    "category": [],
+    "image": "/default/hike-default.png",
+    "kuid": "f9e5f0t5-a8a5-4arb-bfu1-50d392n4",
+    "hikeVersion": "1.0",
+    "weight": "2"
+  }
+}
