+++ src/components/HikeComposer/ExportModal.js
@@ -1,0 +1,33 @@
+import React from "react";
+import { Modal, Button } from "antd";
+
+export default class ExportModal extends React.Component {
+  state = {
+    modal2Visible: false,
+  };
+
+  setModal2Visible(modal2Visible) {
+    this.setState({ modal2Visible });
+  }
+
+  render() {
+    return (
+      <div>
+        <Button type="primary" onClick={() => this.setModal2Visible(true)}>
+          Generate Hike
+        </Button>
+        <Modal
+          title="Compose Successfully"
+          centered
+          visible={this.state.modal2Visible}
+          onOk={() => this.setModal2Visible(false)}
+          onCancel={() => this.setModal2Visible(false)}
+        >
+          <p>some contents...</p>
+          <p>some contents...</p>
+          <p>some contents...</p>
+        </Modal>
+      </div>
+    );
+  }
+}
